pyproject.toml
[project]
name = "moddis-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "dnspython==2.8.0",
  "fastapi==0.135.1",
  "pydantic==2.12.5",
  "pymongo==4.16.0",
  "python-dotenv==1.2.2",
  "uvicorn[standard]==0.42.0",
]
